face-recogn/requirements.txt
opencv-contrib-python
sklearn
pillow
numpy
matplotlib
facenet_pytorch